Track direction using two pins from rotary encoders.
Second encoder pin is not connected anywhere, so I had to solder it
manually. The only free pins left were a3,a6 and a7. But a6 and a7 are
only analog and HAL doesn't allow to use them as normal pins.

This repurposes d5 pin that supposed to be used for HC-SR04 distance meter. But I don't need distance meter at the moment.

I tried using d11 from a speaker, but it didn't work. I am not sure
exactly why, but suppose that amplifier circuit drains all the current,
so voltage can't be detected on it when encoder fires